Refactor connection statuses
