+++ main_window/data_handlers.py
@@ -35,6 +35,8 @@ def process_data(self: "MainWindow", header: packet_spec.PacketHeader, message: 
         case packet_spec.TelemetryPacketSubType.CONTINUITY:
             self.update_continuity_state(message)
             if reset_heartbeat: self.reset_heartbeat_timeout()
+        case packet_spec.TelemetryPacketSubType.CONN_STATUS:
+            pass
         case _:
             pass  
 
@@ -100,8 +102,28 @@ def update_arming_state(self: "MainWindow", message: packet_spec.ArmingStatePack
             self.ui.armingStateValueLabel.setText("5 - ARMED LAUNCH")
             
     self.write_to_log(f"Arming state updated to {message.state}")
-    
-            
+
+def update_pad_server_display(self, status: packet_spec.IPConnectionStatus):
+    match status:
+        case packet_spec.IPConnectionStatus.CONNECTED:
+            self.ui.udpConnStatusLabel.setText("Connected")
+            self.ui.udpConnStatusLabel.setStyleSheet("background-color: rgb(0, 255, 0);")
+        case packet_spec.IPConnectionStatus.CONNECTION_LOST:
+            self.ui.udpConnStatusLabel.setText("Connection lost")
+            self.ui.udpConnStatusLabel.setStyleSheet("background-color: rgb(255, 80, 80);")
+        case packet_spec.IPConnectionStatus.NOT_CONNECTED:
+            self.ui.udpConnStatusLabel.setText("Not connected")
+            self.ui.udpConnStatusLabel.setStyleSheet("background-color: rgb(0, 85, 127);")
+
+def update_serial_connection_display(self, status: packet_spec.SerialConnectionStatus):
+    match status:
+        case packet_spec.SerialConnectionStatus.CONNECTED:                
+            self.ui.serialConnStatusLabel.setText("Connected")
+            self.ui.serialConnStatusLabel.setStyleSheet("background-color: rgb(0, 255, 0);")
+        case packet_spec.SerialConnectionStatus.NOT_CONNECTED:
+            self.ui.serialConnStatusLabel.setText("Not connected")
+            self.ui.serialConnStatusLabel.setStyleSheet("background-color: rgb(0, 85, 127);")
+
 def update_act_state(self: "MainWindow", message: packet_spec.ActuatorStatePacket):
     match message.state:
         case packet_spec.ActuatorState.OFF:
@@ -135,16 +157,17 @@ def filter_data(self: "MainWindow"):
         self.plots[key].points = self.plots[key].points[-(self.graph_range - 1):]
         self.plots[key].data_line.setData(self.plots[key].points)
 
+# Heartbeats only supported for IP connections
 def reset_heartbeat_timeout(self: "MainWindow"):
     self.heartbeat_mutex.lock()
     self.heartbeat_timeout = 10
-    self.update_udp_connection_display(self.UDPConnectionStatus.CONNECTED)
+    self.update_pad_server_display(packet_spec.IPConnectionStatus.CONNECTED)
     self.heartbeat_mutex.unlock()
 
 def decrease_heartbeat(self: "MainWindow"):
     self.heartbeat_mutex.lock()
     self.heartbeat_timeout -= 1
     if self.heartbeat_timeout <= 0:
-        self.update_udp_connection_display(self.UDPConnectionStatus.CONNECTION_LOST)
+        self.update_pad_server_display(packet_spec.IPConnectionStatus.CONNECTION_LOST)
         self.write_to_log(f"Heartbeat not found for {abs(self.heartbeat_timeout) + 1} seconds")
     self.heartbeat_mutex.unlock()
